Session added

+++ customer_api.php
@@ -1,4 +1,7 @@
 <?php
+    session_start();
+    if(isset($_SESSION["login"]))
+    {
     $request_method = $_SERVER['REQUEST_METHOD'];
     switch($request_method)
     {
@@ -10,6 +13,7 @@
             break;
         default:
             echo "Invalid Request";
+    }
     }
 
     function getData()
@@ -43,6 +47,8 @@
       }
       else
       {
+	if($_SESSION["login"]==1)	
+	{
         include 'connectDB.php';
 
         $sql = "INSERT INTO customer VALUES (null,'".$_POST["customer_name"]."','".$_POST["company_name"]."','".$_POST["address"]."','".$_POST["phone"]."','".$_POST["email"]."')";
@@ -52,10 +58,15 @@
           header("Location: customer.php");
         }
         mysqli_close($conn);
-      }
+}
+	else
+	echo "Permission Denied";      
+}
     }
     function putData()
     {
+	if($_SESSION["login"]==1)	
+	{
       include 'connectDB.php';
 
       $sql = "UPDATE customer SET customer_name='".$_POST["customer_name"]."',company_name='".$_POST["company_name"]."',address='".$_POST["address"]."',phone='".$_POST["phone"]."',email='".$_POST["email"]."' WHERE CID='".$_POST["cid"]."';";
@@ -65,9 +76,15 @@
         header("Location: customer.php");
       }
       mysqli_close($conn);
+	}
+	else
+	echo "Permission Denied";
     }
+
     function deleteData()
     {
+if($_SESSION["login"]==1)	
+	{
       include 'connectDB.php';
 
       $sql = "DELETE FROM customer WHERE CID=".$_POST["cid"].";";
@@ -78,4 +95,7 @@
       }
       mysqli_close($conn);
     }
+else
+echo "Permission Denied";
+}
 ?>
